chore: schema 변경 및 초기 데이터 추가

+++ BE/airbnb/src/main/resources/schema.sql
@@ -25,7 +25,7 @@ CREATE TABLE IF NOT EXISTS `airbnb_db`.`accommodation` (
   `type` VARCHAR(45) NOT NULL,
   `num_of_bed` INT NOT NULL DEFAULT 1,
   `num_of_bathroom` INT NOT NULL DEFAULT 1,
-  `price` INT NOT NULL,
+  `price` DECIMAL NOT NULL,
   `address` VARCHAR(45) NOT NULL,
   PRIMARY KEY (`id`))
 ENGINE = InnoDB;
